+++ tests/bas_metadata_library/test_standard_iso_19115_1.py
@@ -537,6 +537,43 @@ def test_identification_maintenance(get_record_response, config_name):
     assert_maintenance(element=identification_maintenance_elements[0], config=config["identification"]["maintenance"])
 
 
+@pytest.mark.usefixtures("get_record_response")
+@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+def test_graphic_overviews(get_record_response, config_name):
+    record = get_record_response(standard=standard, config=config_name)
+    config = configs_safe_v2[config_name]
+
+    if "identification" not in config or "graphic_overviews" not in config["identification"]:
+        pytest.skip("record does not contain graphic overviews")
+
+    for graphic_overview in config["identification"]["graphic_overviews"]:
+        graphic_overview_elements = record.xpath(
+            f"/gmd:MD_Metadata/gmd:identificationInfo/gmd:MD_DataIdentification/gmd:graphicOverview[@id='{graphic_overview['identifier']}']",
+            namespaces=namespaces.nsmap(),
+        )
+        assert len(graphic_overview_elements) == 1
+
+        href_element = graphic_overview_elements[0].xpath(
+            f"./gmd:MD_BrowseGraphic/gmd:fileName/gco:CharacterString/text() = '{graphic_overview['href']}'",
+            namespaces=namespaces.nsmap(),
+        )
+        assert href_element is True
+
+        if "description" in graphic_overview:
+            description_element = graphic_overview_elements[0].xpath(
+                f"./gmd:MD_BrowseGraphic/gmd:fileDescription/gco:CharacterString/text() = '{graphic_overview['description']}'",
+                namespaces=namespaces.nsmap(),
+            )
+            assert description_element is True
+
+        if "mime_type" in graphic_overview:
+            mime_type_element = graphic_overview_elements[0].xpath(
+                f"./gmd:MD_BrowseGraphic/gmd:fileType/gco:CharacterString/text() = '{graphic_overview['mime_type']}'",
+                namespaces=namespaces.nsmap(),
+            )
+            assert mime_type_element is True
+
+
 def _resolve_descriptive_keywords_xpaths(config) -> List[dict]:
     keywords = []
     for keyword in config: